Organize coordinator directory.
