+++ project3/coordinator/coordinator_resources/connected_participants.cpp
@@ -2,6 +2,7 @@
  * @file Implementation of ConnectedParticipants class.
  */
 #include "connected_participants.h"
+
 #include <loguru.hpp>
 namespace coordinator {
