+++ project3/coordinator/coordinator_resources/connected_participants.h
