+++ project3/coordinator/coordinator_resources/message_log.cpp
