+++ project3/coordinator/coordinator_resources/registrar.cpp
